\name{fSNP}
\alias{fSNP}
\docType{data}
\title{
Example data of negative SNP
}
\description{
low quality snp data
}
\usage{data("fSNP")}
\format{
  The format is:
 chr [1:1031] "0500021626C" "0500115741G" ...
}
\source{
http://www.ncpgr.cn/supplements/MPR_genotyping/MPR_genotyping.tar.gz
}
\examples{
#load data
data(fSNP)
}
\keyword{datasets}
