\name{makeEmissionFUN}
\alias{makeEmissionFUN}
\title{
emission of HMM
}
\description{
make emission matrix for HMM
}
\usage{
makeEmissionFUN(errorRate = 0.01)
}
\arguments{
  \item{errorRate}{
the error rate of RIL genotype calls at a specific SNP site
}
}
\seealso{
\code{\link{hmm.vitFUN.rils}}
}
\keyword{ HMM }
