\name{markerData}
\alias{markerData}
\docType{data}
\title{
Data of Markers
}
\description{
Load overlapping SNP alleles from low-depth sequences of one parent,
you can also use genotype results from traditional genotyping studies,
e.g. results of SSR/RFLP markers.
}
\usage{data("markerData")}
\format{
  The format is:
 Named chr [1:272] "G" "T" "A" "C" "G" "C" "A" "T" "A" "C" "T" "A" ...
 - attr(*, "names")= chr [1:272] "0500033821G" "0500059721T" "0500125700G" "0500169119C" ...
}
\source{
http://www.ncpgr.cn/supplements/MPR_genotyping/MPR_genotyping.tar.gz
}
\examples{
data(markerData)
}
\keyword{datasets}
