\name{phy2get.haldane.rils}
\alias{phy2get.haldane.rils}
\title{
transition of HMM
}
\description{
make matrix of transition for HMM
}
\usage{
phy2get.haldane.rils(a, b, l)
}
\arguments{
  \item{a}{
state now
}
  \item{b}{
state next
}
  \item{l}{
the distance between two positions
}
}
\keyword{ HMM }
